# api/openapi/inventory.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Logistics Management inventory
  description: TODO
  contact:
    name: Mousa Zeydan
    email: mous.zeydan@gmail.com
servers:
  - url: https://{hostname}/api
    variables:
      hostname:
        default: localhost
#security:
#  - bearerAuth: []
paths:
  /locations:
    post:
      operationId: createLocation
      requestBody:
        description: todo
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '204':
          description: todo
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: getLocations
      responses:
        '200':
          description: todo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Locations'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /locations/{locationUUID}:
    get:
      operationId: getLocationByUUID
      parameters:
        - in: path
          name: locationUUID
          schema:
            type: string
            format: uuid
          required: true
          description: todo
      responses:
        '200':
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
        operationId: updateLocation
        parameters:
          - in: path
            name: locationUUID
            schema:
                type: string
                format: uuid
            required: true
            description: todo
        requestBody:
            description: todo
            required: true
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Location'
        responses:
            '204':
              description: todo
            default:
              description: unexpected error
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/Error'
  /locations/{locationUUID}/products:
    get:
      operationId: getLocationContents
      parameters:
        - in: path
          name: locationUUID
          schema:
            type: string
            format: uuid
          required: true
          description: todo
      responses:
        '200':
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductStocks'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
        operationId: addProductsToLocation
        parameters:
            - in: path
              name: locationUUID
              schema:
                type: string
                format: uuid
              required: true
              description: todo
        requestBody:
            description: todo
            required: true
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    required: [ productUUID, quantity ]
                    properties:
                      productUUID:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
        responses:
          '204':
            description: todo
          default:
            description: unexpected error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
  /locations/{locationUUID}/products/{productUUID}/report-damage:
    put:
      operationId: reportDamagedProduct
      parameters:
        - in: path
          name: locationUUID
          schema:
            type: string
            format: uuid
          required: true
          description: todo
        - in: path
          name: productUUID
          schema:
            type: string
            format: uuid
          required: true
          description: todo
      requestBody:
        description: todo
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ reason, quantity ]
              properties:
                reason:
                  type: string
                quantity:
                  type: integer
      responses:
        '204':
          description: todo
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /products:
    post:
      operationId: createProduct
      requestBody:
        description: todo
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '204':
          description: todo
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
        operationId: getProducts
        responses:
            '200':
              description: todo
              content:
                application/json:
                  schema:
                    type: array
                    items:
                      $ref: '#/components/schemas/Products'
            default:
              description: unexpected error
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/Error'
  /products/{productUUID}:
    get:
      operationId: getProductByUUID
      parameters:
        - in: path
          name: productUUID
          schema:
            type: string
            format: uuid
          required: true
          description: todo
      responses:
        '200':
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
        operationId: updateProduct
        parameters:
          - in: path
            name: productUUID
            schema:
                type: string
                format: uuid
            required: true
            description: todo
        requestBody:
            description: todo
            required: true
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Product'
        responses:
            '204':
              description: todo
            default:
              description: unexpected error
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/Error'
  /product/{productUUID}/locations:
    get:
      operationId: getProductLocations
      parameters:
        - in: path
          name: productUUID
          schema:
            type: string
            format: uuid
          required: true
          description: todo
      responses:
        '200':
          description: todo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductLocationInventory'

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /products/transfer:
    put:
      operationId: transferProducts
      requestBody:
        description: todo
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferProductRequest'
      responses:
        '204':
          description: todo
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


components:
  #  securitySchemes:
  #    bearerAuth:
  #      type: http
  #      scheme: bearer
  #      bearerFormat: JWT
  schemas:
    Product:
      type: object
      required: [name, price, weight]
      properties:
        productUUID:
          type: string
        name:
          type: string
        weight:
          type: number
        price:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    TransferProductRequest:
      type: object
      required: [ sourceLocationUUID, destLocationUUID ,productUUID, quantity ]
      properties:
        sourceLocationUUID:
          type: string
          format: uuid
        destLocationUUID:
          type: string
          format: uuid
        productUUID:
          type: string
          format: uuid
        quantity:
          type: integer

    ProductStock:
      type: object
      required: [ name, price, weight, quantity ]
      properties:
        productUUID:
          type: string
        name:
          type: string
        quantity:
          type: integer
        weight:
          type: number
        price:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time


    Location:
      type: object
      required: [name, city, address]
      properties:
        locationUUID:
          type: string
        name:
          type: string
        city:
          type: string
        address:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProductLocationInventory:
      description: A location along with the stock quantity for a specific product
      allOf:
        - $ref: '#/components/schemas/Location'
        - type: object
          properties:
            availableQuantity:
              type: integer
              description: The quantity of the product available at this location
              example: 42
            damagedQuantity:
              type: integer
              description: The quantity of the product damaged at this location
              example: 10
            reservedQuantity:
              type: integer
              description: The quantity of the product reserved at this location
              example: 15
    ProductLocationInventories:
      type: object
      required: [ items ]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductLocationInventory'

    Products:
      type: object
      required: [ products ]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'

    ProductStocks:
      type: object
      required: [ product_stocks ]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductStock'

    Locations:
      type: object
      required: [ locations ]
      properties:
        locations:
          type: array
          items:
            $ref: '#/components/schemas/Location'

    Error:
      type: object
      required:
        - slug
        - message
      properties:
        slug:
          type: string
        message:
          type: string