# internal/inventory/ports/openapi_api.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package ports

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /locations)
	GetLocations(w http.ResponseWriter, r *http.Request)

	// (POST /locations)
	CreateLocation(w http.ResponseWriter, r *http.Request)

	// (GET /locations/{locationUUID})
	GetLocationByUUID(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID)

	// (PUT /locations/{locationUUID})
	UpdateLocation(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID)

	// (GET /locations/{locationUUID}/products)
	GetLocationContents(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID)

	// (PUT /locations/{locationUUID}/products)
	AddProductsToLocation(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID)

	// (PUT /locations/{locationUUID}/products/{productUUID}/report-damage)
	ReportDamagedProduct(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID, productUUID openapi_types.UUID)

	// (GET /product/{productUUID}/locations)
	GetProductLocations(w http.ResponseWriter, r *http.Request, productUUID openapi_types.UUID)

	// (GET /products)
	GetProducts(w http.ResponseWriter, r *http.Request)

	// (POST /products)
	CreateProduct(w http.ResponseWriter, r *http.Request)

	// (PUT /products/transfer)
	TransferProducts(w http.ResponseWriter, r *http.Request)

	// (GET /products/{productUUID})
	GetProductByUUID(w http.ResponseWriter, r *http.Request, productUUID openapi_types.UUID)

	// (PUT /products/{productUUID})
	UpdateProduct(w http.ResponseWriter, r *http.Request, productUUID openapi_types.UUID)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (GET /locations)
func (_ Unimplemented) GetLocations(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /locations)
func (_ Unimplemented) CreateLocation(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /locations/{locationUUID})
func (_ Unimplemented) GetLocationByUUID(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (PUT /locations/{locationUUID})
func (_ Unimplemented) UpdateLocation(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /locations/{locationUUID}/products)
func (_ Unimplemented) GetLocationContents(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (PUT /locations/{locationUUID}/products)
func (_ Unimplemented) AddProductsToLocation(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (PUT /locations/{locationUUID}/products/{productUUID}/report-damage)
func (_ Unimplemented) ReportDamagedProduct(w http.ResponseWriter, r *http.Request, locationUUID openapi_types.UUID, productUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /product/{productUUID}/locations)
func (_ Unimplemented) GetProductLocations(w http.ResponseWriter, r *http.Request, productUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /products)
func (_ Unimplemented) GetProducts(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /products)
func (_ Unimplemented) CreateProduct(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (PUT /products/transfer)
func (_ Unimplemented) TransferProducts(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /products/{productUUID})
func (_ Unimplemented) GetProductByUUID(w http.ResponseWriter, r *http.Request, productUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (PUT /products/{productUUID})
func (_ Unimplemented) UpdateProduct(w http.ResponseWriter, r *http.Request, productUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetLocations operation middleware
func (siw *ServerInterfaceWrapper) GetLocations(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetLocations(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateLocation operation middleware
func (siw *ServerInterfaceWrapper) CreateLocation(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateLocation(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetLocationByUUID operation middleware
func (siw *ServerInterfaceWrapper) GetLocationByUUID(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "locationUUID" -------------
	var locationUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "locationUUID", chi.URLParam(r, "locationUUID"), &locationUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "locationUUID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetLocationByUUID(w, r, locationUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateLocation operation middleware
func (siw *ServerInterfaceWrapper) UpdateLocation(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "locationUUID" -------------
	var locationUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "locationUUID", chi.URLParam(r, "locationUUID"), &locationUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "locationUUID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateLocation(w, r, locationUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetLocationContents operation middleware
func (siw *ServerInterfaceWrapper) GetLocationContents(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "locationUUID" -------------
	var locationUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "locationUUID", chi.URLParam(r, "locationUUID"), &locationUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "locationUUID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetLocationContents(w, r, locationUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AddProductsToLocation operation middleware
func (siw *ServerInterfaceWrapper) AddProductsToLocation(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "locationUUID" -------------
	var locationUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "locationUUID", chi.URLParam(r, "locationUUID"), &locationUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "locationUUID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AddProductsToLocation(w, r, locationUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ReportDamagedProduct operation middleware
func (siw *ServerInterfaceWrapper) ReportDamagedProduct(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "locationUUID" -------------
	var locationUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "locationUUID", chi.URLParam(r, "locationUUID"), &locationUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "locationUUID", Err: err})
		return
	}

	// ------------- Path parameter "productUUID" -------------
	var productUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "productUUID", chi.URLParam(r, "productUUID"), &productUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "productUUID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ReportDamagedProduct(w, r, locationUUID, productUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetProductLocations operation middleware
func (siw *ServerInterfaceWrapper) GetProductLocations(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "productUUID" -------------
	var productUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "productUUID", chi.URLParam(r, "productUUID"), &productUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "productUUID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetProductLocations(w, r, productUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetProducts operation middleware
func (siw *ServerInterfaceWrapper) GetProducts(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetProducts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateProduct operation middleware
func (siw *ServerInterfaceWrapper) CreateProduct(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateProduct(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// TransferProducts operation middleware
func (siw *ServerInterfaceWrapper) TransferProducts(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.TransferProducts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetProductByUUID operation middleware
func (siw *ServerInterfaceWrapper) GetProductByUUID(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "productUUID" -------------
	var productUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "productUUID", chi.URLParam(r, "productUUID"), &productUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "productUUID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetProductByUUID(w, r, productUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateProduct operation middleware
func (siw *ServerInterfaceWrapper) UpdateProduct(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "productUUID" -------------
	var productUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "productUUID", chi.URLParam(r, "productUUID"), &productUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "productUUID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateProduct(w, r, productUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/locations", wrapper.GetLocations)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/locations", wrapper.CreateLocation)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/locations/{locationUUID}", wrapper.GetLocationByUUID)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/locations/{locationUUID}", wrapper.UpdateLocation)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/locations/{locationUUID}/products", wrapper.GetLocationContents)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/locations/{locationUUID}/products", wrapper.AddProductsToLocation)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/locations/{locationUUID}/products/{productUUID}/report-damage", wrapper.ReportDamagedProduct)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/product/{productUUID}/locations", wrapper.GetProductLocations)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/products", wrapper.GetProducts)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/products", wrapper.CreateProduct)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/products/transfer", wrapper.TransferProducts)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/products/{productUUID}", wrapper.GetProductByUUID)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/products/{productUUID}", wrapper.UpdateProduct)
	})

	return r
}