# internal/inventory/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package ports

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// Location defines model for Location.
type Location struct {
	Address      string     `json:"address"`
	City         string     `json:"city"`
	CreatedAt    *time.Time `json:"createdAt,omitempty"`
	LocationUUID *string    `json:"locationUUID,omitempty"`
	Name         string     `json:"name"`
	UpdatedAt    *time.Time `json:"updatedAt,omitempty"`
}

// Locations defines model for Locations.
type Locations struct {
	Locations []Location `json:"locations"`
}

// Product defines model for Product.
type Product struct {
	CreatedAt   *time.Time `json:"createdAt,omitempty"`
	Name        string     `json:"name"`
	Price       float32    `json:"price"`
	ProductUUID *string    `json:"productUUID,omitempty"`
	UpdatedAt   *time.Time `json:"updatedAt,omitempty"`
	Weight      float32    `json:"weight"`
}

// ProductLocationInventory defines model for ProductLocationInventory.
type ProductLocationInventory struct {
	Address string `json:"address"`

	// AvailableQuantity The quantity of the product available at this location
	AvailableQuantity *int       `json:"availableQuantity,omitempty"`
	City              string     `json:"city"`
	CreatedAt         *time.Time `json:"createdAt,omitempty"`

	// DamagedQuantity The quantity of the product damaged at this location
	DamagedQuantity *int    `json:"damagedQuantity,omitempty"`
	LocationUUID    *string `json:"locationUUID,omitempty"`
	Name            string  `json:"name"`

	// ReservedQuantity The quantity of the product reserved at this location
	ReservedQuantity *int       `json:"reservedQuantity,omitempty"`
	UpdatedAt        *time.Time `json:"updatedAt,omitempty"`
}

// ProductStock defines model for ProductStock.
type ProductStock struct {
	CreatedAt   *time.Time `json:"createdAt,omitempty"`
	Name        string     `json:"name"`
	Price       float32    `json:"price"`
	ProductUUID *string    `json:"productUUID,omitempty"`
	Quantity    int        `json:"quantity"`
	UpdatedAt   *time.Time `json:"updatedAt,omitempty"`
	Weight      float32    `json:"weight"`
}

// ProductStocks defines model for ProductStocks.
type ProductStocks struct {
	Items *[]ProductStock `json:"items,omitempty"`
}

// Products defines model for Products.
type Products struct {
	Items *[]Product `json:"items,omitempty"`
}

// TransferProductRequest defines model for TransferProductRequest.
type TransferProductRequest struct {
	DestLocationUUID   openapi_types.UUID `json:"destLocationUUID"`
	ProductUUID        openapi_types.UUID `json:"productUUID"`
	Quantity           int                `json:"quantity"`
	SourceLocationUUID openapi_types.UUID `json:"sourceLocationUUID"`
}

// AddProductsToLocationJSONBody defines parameters for AddProductsToLocation.
type AddProductsToLocationJSONBody = []struct {
	ProductUUID openapi_types.UUID `json:"productUUID"`
	Quantity    int                `json:"quantity"`
}

// ReportDamagedProductJSONBody defines parameters for ReportDamagedProduct.
type ReportDamagedProductJSONBody struct {
	Quantity int    `json:"quantity"`
	Reason   string `json:"reason"`
}

// CreateLocationJSONRequestBody defines body for CreateLocation for application/json ContentType.
type CreateLocationJSONRequestBody = Location

// UpdateLocationJSONRequestBody defines body for UpdateLocation for application/json ContentType.
type UpdateLocationJSONRequestBody = Location

// AddProductsToLocationJSONRequestBody defines body for AddProductsToLocation for application/json ContentType.
type AddProductsToLocationJSONRequestBody = AddProductsToLocationJSONBody

// ReportDamagedProductJSONRequestBody defines body for ReportDamagedProduct for application/json ContentType.
type ReportDamagedProductJSONRequestBody ReportDamagedProductJSONBody

// CreateProductJSONRequestBody defines body for CreateProduct for application/json ContentType.
type CreateProductJSONRequestBody = Product

// TransferProductsJSONRequestBody defines body for TransferProducts for application/json ContentType.
type TransferProductsJSONRequestBody = TransferProductRequest

// UpdateProductJSONRequestBody defines body for UpdateProduct for application/json ContentType.
type UpdateProductJSONRequestBody = Product